Rename codegen files because the collision with reflex/schema.h was getting annoying
